Swapping Turrets
The ability to swap control of the turrets is implemented. By pressing mouse1, you can select a turret to control. Space to fire. (Still To Do: make it so collision with the player and the pressing of a key switches control instead. (Also removed some unused folders/ moved some contents into folders)

Co-Authored-By: S41NT <82161795+AleksanderBabii@users.noreply.github.com>